git-svn-id: http://svn.ilias.de/svn/ilias/branches/Release_4_4_x_branch@51784 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/Test/classes/class.ilTestExportGUI.php
@@ -247,8 +247,16 @@ class ilTestExportGUI extends ilExportGUI
 		$export_dir = $this->obj->getExportDirectory();
 		foreach($_POST['file'] as $file)
 		{
+			$file = basename($file);
+			$dir = substr($file, 0, strlen($file) - 4);
+
+			if( !strlen($file) || !strlen($dir) )
+			{
+				continue;
+			}
+			
 			$exp_file = $export_dir.'/'.$file;
-			$exp_dir = $export_dir.'/'.substr($file, 0, strlen($file) - 4);
+			$exp_dir = $export_dir.'/'.$dir;
 			if(@is_file($exp_file))
 			{
 				unlink($exp_file);
